GUI Button V1
So this is the simple GUI I made with a few buttons...they don't do
anything, but that will be updated.  I'm not actually sure that I added
this class to our src correctly.  I had made a project on my local
computer, and I copied the package and class from that into this branch.
Lmk what you guys think.